requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
pyyaml>=6.0.0
scikit-learn>=1.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0
mypy>=1.5.0

